Update dex structure to match module standard
