+++ contracts/modules/extensions/dex/src/exchanges/exchange_resolver.rs
@@ -1,0 +1,48 @@
+use crate::DEX;
+use crate::dex_trait::Identify;
+use crate::error::DexError;
+
+// Supported exchanges on Juno
+#[cfg(feature = "juno")]
+pub use crate::exchanges::junoswap::{JunoSwap, JUNOSWAP};
+
+#[cfg(any(feature = "juno", feature = "terra"))]
+pub use crate::exchanges::loop_dex::{Loop, LOOP};
+
+#[cfg(feature = "terra")]
+pub use crate::exchanges::terraswap::{Terraswap, TERRASWAP};
+
+#[cfg(any(feature = "juno", feature = "osmosis"))]
+pub use crate::exchanges::osmosis::{Osmosis, OSMOSIS};
+
+pub(crate) fn identify_exchange(value: &str) -> Result<&'static dyn Identify, DexError> {
+    match value {
+        #[cfg(feature = "juno")]
+        JUNOSWAP => Ok(&JunoSwap {}),
+        #[cfg(feature = "juno")]
+        OSMOSIS => Ok(&Osmosis {
+            local_proxy_addr: None,
+        }),
+        #[cfg(any(feature = "juno", feature = "terra"))]
+        LOOP => Ok(&Loop {}),
+        #[cfg(feature = "terra")]
+        TERRASWAP => Ok(&Terraswap {}),
+        _ => Err(DexError::UnknownDex(value.to_owned())),
+    }
+}
+
+pub(crate) fn resolve_exchange(value: &str) -> Result<&'static dyn DEX, DexError> {
+    match value {
+        #[cfg(feature = "juno")]
+        JUNOSWAP => Ok(&JunoSwap {}),
+        // #[cfg(feature = "osmosis")]
+        // OSMOSIS => Ok(&Osmosis {
+        //     local_proxy_addr: None,
+        // }),
+        #[cfg(any(feature = "juno", feature = "terra"))]
+        LOOP => Ok(&Loop {}),
+        #[cfg(feature = "terra")]
+        TERRASWAP => Ok(&Terraswap {}),
+        _ => Err(DexError::ForeignDex(value.to_owned())),
+    }
+}
