+++ contracts/modules/extensions/dex/src/exchanges/mod.rs
@@ -7,3 +7,6 @@ pub mod terraswap;
 
 #[cfg(any(feature = "juno", feature = "osmosis"))]
 pub mod osmosis;
+
+
+pub(crate) mod exchange_resolver;
