+++ contracts/modules/extensions/dex/src/handlers/mod.rs
@@ -1,0 +1,4 @@
+mod execute;
+mod query;
+
+pub use {execute::execute_handler, query::query_handler};
