+++ contracts/modules/extensions/dex/src/lib.rs
@@ -3,7 +3,8 @@ pub mod contract;
 pub(crate) mod dex_trait;
 pub mod error;
 mod exchanges;
-pub(crate) mod queries;
+
+pub(crate) mod handlers;
 
 pub use commands::LocalDex;
 pub use dex_trait::DEX;
